+++ albums.html
@@ -9,7 +9,11 @@
 </head>
 <body>
     <div id="albums-wrapper"></div>
-    <script src="albums.js"></script>
-    <script src="header.js"></script>
+    
+    <script type="module" src="./albums.js"></script>
+    <script type="module" src="./header.js"></script>
+    <script type="module" src="./export.js"></script>
+    
+    <!-- <script src="./functions.js"></script> -->
 </body>
 </html>
