+++ albums.js
@@ -1,13 +1,9 @@
-let albumsElement = document.createElement('div')
-let postAlbumElementsTitle = document.createElement('h2')
-postAlbumElementsTitle.classList.add('post-album-elements-title')
+import {toUpperCase, renderSingleAlbum, renderAllAlbums} from "./export.js"
 
 
 let queryParams = document.location.search;
 let urlParams = new URLSearchParams(queryParams);
 let userId = urlParams.get('user_id');
-
-
 
 function init() {
  if(userId){
@@ -18,31 +14,16 @@ function init() {
 }
 
 
-function renderAllAlbums() {
-  fetch('https://jsonplaceholder.typicode.com/albums?_expand=user&_embed=photos&_limit=15')
-    .then(res => res.json())
-    .then(albums => {
-      console.log(albums);
-      albums.map(singleAlbum => {
-        renderSingleAlbum({
-          album: singleAlbum,
-          title: 'All albums:',
-          createdBy: `<div>Album created by: <a href="./user.html?user_id=${singleAlbum.user.id}">${singleAlbum.user.name}</a></div>`,
-        });
-      })
-    })
-  }
 
-
-function renderAlbumsByUserId(id) {
-   
+function renderAlbumsByUserId(id) {  
     fetch(`https://jsonplaceholder.typicode.com/users/${id}/albums?_embed=photos&_expand=user`)
       .then(res => res.json())
       .then(albums => {
+        
         albums.map(singleAlbum => {
           let albumData = {
             album: singleAlbum,
-            title: `Albums of ${singleAlbum.user.name}:`,
+            title: `Albums of ${toUpperCase(singleAlbum.user.name)}:`,
             createdBy: `${singleAlbum.user.name}`,
           }
   
@@ -53,22 +34,22 @@ function renderAlbumsByUserId(id) {
   
 init()
 
-function renderSingleAlbum(data) {
+// function renderSingleAlbum(data) {
 
-  let {album, title, createdBy} = data;
+//   let {album, title, createdBy} = data;
 
-  let albumItem = document.createElement('div');
-  albumItem.classList.add('album-item');
+//   let albumItem = document.createElement('div');
+//   albumItem.classList.add('album-item');
   
-  postAlbumElementsTitle.textContent = title;
+//   postAlbumElementsTitle.textContent = title;
 
-  let randomIndex = Math.floor(Math.random() * album.photos.length);
+//   let randomIndex = Math.floor(Math.random() * album.photos.length);
 
-  albumItem.innerHTML = `<h3><a href="./album.html?album_id=${album.id}&album_title=${album.title}&user_id=${album.userId}&user_name=${album.user.name}">${album.title}</a> (${album.photos.length})</h3>
-                        ${createdBy}
-                        <img src="${album.photos[randomIndex].thumbnailUrl}">`;
+//   albumItem.innerHTML = `<h3><a href="./album.html?album_id=${album.id}&album_title=${album.title}&user_id=${album.userId}&user_name=${album.user.name}">${toUpperCase(album.title)}</a> (${album.photos.length})</h3>
+//                         ${createdBy}
+//                         <img src="${album.photos[randomIndex].thumbnailUrl}">`;
 
-                        albumsElement.prepend(albumItem); 
-                        postAlbumElementsTitle.innerHTML = `album of ${createdBy}`
-                        document.body.append(albumsElement)
-}
+//                         albumsElement.prepend(albumItem); 
+//                         postAlbumElementsTitle.innerHTML = `album of ${createdBy}`
+//                         document.body.append(albumsElement)
+// }
