+++ index.html
@@ -13,8 +13,8 @@
     <div id="post-wrapper"></div>
     <div id="album-wrapper"></div>
     
-    <script src="script.js"></script>
-    <script src="header.js"></script>
-    <script src="functions.js"></script>
+    <script type="module" src="script.js"></script>
+    <script type="module" src="header.js"></script>
+    <script type="module" src="functions.js"></script>
 </body>
 </html>
