+++ user.css
@@ -1,0 +1,83 @@
+
+
+
+  #users-wrapper {
+    background-color: #ffc8c821;
+  }
+
+  
+
+
+
+ 
+
+body {
+    background-color: #5e3030;
+    color: lightgrey;
+    text-shadow: 1px 1px black;
+    font-family: 'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif;
+  }
+
+  nav {
+        text-align: center;
+        border: 1px solid lightgray;
+        background: #f2f2f255;
+  } 
+  nav a {
+    margin-left: 6rem;
+    font-size: 30px;
+    font-family: Impact, Haettenschweiler, 'Arial Narrow Bold', sans-serif;
+    text-shadow: 2px 2px black
+  }
+
+  h1 {
+    text-align: center;
+    font-size: 3rem;
+    font-family: Verdana, Geneva, Tahoma, sans-serif;
+    text-shadow: 2px 2px black;
+    border: 1px solid currentColor;
+    padding: 20px;
+  }
+  .post-item {
+    background-color:#ffc8c821;
+    padding: 20px;
+  }
+  #post-wrapper {
+    display: flex;
+    flex-direction: row;
+    width: 100%;
+    justify-content: center;
+    flex-wrap: wrap;
+  }
+  .post-item {
+    width: 40%;
+    margin: 20px;
+  }
+  .post-item a {
+    margin: 20px;
+    padding: 20px;
+    text-align: center;
+  }
+  form {
+    text-align: right;
+    margin: 20px;
+  }
+
+  .albums-element {
+    display: flex;
+    width: 100%;
+    flex-wrap: wrap;
+    flex-direction: row;
+    justify-content: center;
+  }
+
+  .albums-element img {
+    width:10rem;
+    height:10rem;
+    justify-content: center;
+  }
+  .albums-element h3 {
+    width: 5rem;
+    justify-content: center;
+    text-align: center;
+  }
