+++ user.html
@@ -5,7 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>User page</title>
-    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="user.css">
 </head>
 <body>
