+++ users.html
@@ -10,8 +10,9 @@
 <body>
 
     <div id="users-wrapper"></div>
-    <script src="users.js"></script>    
-    <script src="header.js"></script>
-    <script src="functions.js"></script>
+    <script  type="module"  src="users.js"></script>    
+    <script type="module" src="header.js"></script>
+
+    <script type="module" src="./export.js"></script>
 </body>
 </html>
